more updates to components
