+++ src/App.js
@@ -1,15 +1,18 @@
 import React from 'react';
 
 import { connect } from "react-redux";
-import {addItem} from "./actions/addItem";
+//actions
+import {addItem} from "./actions/action";
 import {removeItem} from "./actions/removeItem";
-
+//components
 import Header from './components/Header';
 import AddedFeatures from './components/AddedFeatures';
 import AdditionalFeatures from './components/AdditionalFeatures';
 import Total from './components/Total';
 
 const App = props=> {
+  console.log("app.js", props);
+  //Props contain state, actions for reducer, that are passed in using connect
   
   const removeFeature = item => {
     // dispatch an action here to remove an item
@@ -17,10 +20,10 @@ const App = props=> {
     
   };
 
-  const buyItem = item => {
-    // dipsatch an action here to add an item
-    props.addItem(item)
-  };
+  // const buyItem = item => {
+  //   // dipsatch an action here to add an item
+  //   props.addItem(item)
+  // };
 
   return (
     <div className="boxes">
@@ -30,7 +33,7 @@ const App = props=> {
       </div>
       <div className="box">
         <AdditionalFeatures additionalFeatures={props.additionalFeatures} 
-        buyItem={buyItem}/>
+        buyItem={props.buyItem}/>
         <Total car={props.car} additionalPrice={props.additionalPrice} />
       </div>
     </div>
