--- src/actions/addItem.js
@@ -1,6 +1,0 @@
-export const addItem = item =>{
-    return{
-        type: "ADD_ITEM", payload: {id:item.id, price: item.price}
-    }
-}
-
