+++ src/components/AdditionalFeature.js
@@ -1,10 +1,13 @@
 import React from 'react';
 
+
 const AdditionalFeature = props => {
+  
   return (
     <li>
       {/* Add an onClick that will let you add a feature to your car */}
-      <button className="button">Add</button>
+      <button className="button" 
+      onClick ={()=>props.buyItem(props.feature)}>Add</button>
       {props.feature.name} (+{props.feature.price})
     </li>
   );
