+++ src/components/AdditionalFeatures.js
@@ -1,14 +1,19 @@
 import React from 'react';
 import AdditionalFeature from './AdditionalFeature';
+import {addItem  }from '../actions/action';
 
 const AdditionalFeatures = props => {
+  const buyItem = item => {
+    // dipsatch an action here to add an item
+    props.addItem(item)
+  };
   return (
     <div className="content">
       <h4>Additional Features</h4>
       {props.additionalFeatures.length ? (
         <ol type="1">
           {props.additionalFeatures.map(item => (
-            <AdditionalFeature key={item.id} feature={item} />
+            <AdditionalFeature key={item.id} feature={item} buyItem={buyItem} />
           ))}
         </ol>
       ) : (
